@@ -10,8 +10,8 @@ environment:
 dependencies:
   flutter:
     sdk: flutter
-  dio: ^4.0.6
-  webview_flutter: ^4.0.2
+  dio: ^5.3.4
+  webview_flutter: ^4.4.2
 
 dev_dependencies:
   flutter_test:
